refactored code for moving utils folders (#350)
