+++ public/js/actions/tests/breakpoints.js
@@ -1,8 +1,8 @@
 declare var describe: (name: string, func: () => void) => void;
 declare var it: (desc: string, func: () => void) => void;
 
-const { createStore, selectors, actions } = require("../../util/test-head");
-const { Task } = require("../../util/task");
+const { createStore, selectors, actions } = require("../../utils/test-head");
+const { Task } = require("../../utils/task");
 const expect = require("expect.js");
 
 const simpleMockThreadClient = {
