+++ public/js/clients/chrome.js
@@ -4,7 +4,7 @@ const { connect } = require("../lib/chrome-remote-debug-protocol");
 const defer = require("../lib/devtools/shared/defer");
 const { Tab } = require("../types");
 const { isEnabled, getValue } = require("../../../config/feature");
-const { networkRequest } = require("../util/networkRequest");
+const { networkRequest } = require("../utils/networkRequest");
 const { setupCommands, clientCommands } = require("./chrome/commands");
 const { setupEvents, clientEvents } = require("./chrome/events");
