+++ public/js/components/SourceTabs.js
@@ -5,7 +5,7 @@ const { connect } = require("react-redux");
 const { bindActionCreators } = require("redux");
 const Isvg = React.createFactory(require("react-inlinesvg"));
 const { getSelectedSource, getSourceTabs } = require("../selectors");
-const { endTruncateStr } = require("../util/utils");
+const { endTruncateStr } = require("../utils/utils");
 const classnames = require("classnames");
 const actions = require("../actions");
 const { isEnabled } = require("../../../config/feature");
