+++ public/js/main.js
@@ -7,8 +7,8 @@ const React = require("react");
 
 const DevToolsUtils = require("devtools-sham/shared/DevToolsUtils");
 const AppConstants = require("devtools-sham/sham/appconstants").AppConstants;
-const { injectGlobals } = require("./util/debug");
-const { log } = require("./util/utils");
+const { injectGlobals } = require("./utils/debug");
+const { log } = require("./utils/utils");
 const { isEnabled, isFirefoxPanel, isDevelopment } = require("../../config/feature");
 
 // Set various flags before requiring app code.
@@ -20,7 +20,7 @@ log("Debugger bootstrapping");
 
 const { getClient, connectClients, startDebugging } = require("./clients");
 const firefox = require("./clients/firefox");
-const { configureStore } = require("./util/create-store");
+const { configureStore } = require("./utils/create-store");
 const reducers = require("./reducers");
 
 const Tabs = require("./components/Tabs");
