+++ public/js/reducers/sources.js
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const fromJS = require("../util/fromJS");
+const fromJS = require("../utils/fromJS");
 const I = require("immutable");
 
 import type { Action, Source } from "../actions/types";
